Moved classes to name package and added dependency.
