+++ nammu/src/main/java/pl/tajchert/nammu/support/OnPermissionRequestCallback.java
@@ -1,4 +1,4 @@
-package sixgreen.nammu;
+package pl.tajchert.nammu.support;
 
 import pl.tajchert.nammu.PermissionCallback;
